requirements.txt
django>=1.5.1
wheel==0.23.0
factory-boy==2.4.1
